Refactor redeem voucher into separate module
