+++ android/app/build.gradle.kts
@@ -398,6 +398,7 @@ dependencies {
     implementation(projects.lib.feature.filter.impl)
     implementation(projects.lib.feature.multihop.impl)
     implementation(projects.lib.feature.problemreport.impl)
+    implementation(projects.lib.feature.redeemvoucher.impl)
     implementation(projects.lib.feature.splittunneling.impl)
     implementation(projects.lib.feature.serveripoverride.impl)
     implementation(projects.lib.map)
