--- android/app/src/main/kotlin/net/mullvad/mullvadvpn/constant/CommonConstant.kt
@@ -1,3 +1,0 @@
-package net.mullvad.mullvadvpn.constant
-
-const val VOUCHER_LENGTH = 16
