+++ android/lib/feature/redeemvoucher/impl/src/main/java/net/mullvad/mullvadvpn/feature/redeemvoucher/impl/CommonConstant.kt
@@ -1,0 +1,3 @@
+package net.mullvad.mullvadvpn.feature.redeemvoucher.impl
+
+const val VOUCHER_LENGTH = 16
