+++ android/lib/feature/redeemvoucher/impl/src/main/java/net/mullvad/mullvadvpn/feature/redeemvoucher/impl/RedeemVoucherDialog.kt
@@ -1,4 +1,4 @@
-package net.mullvad.mullvadvpn.compose.dialog
+package net.mullvad.mullvadvpn.feature.redeemvoucher.impl
 
 import android.content.res.Configuration
 import androidx.compose.foundation.Image
@@ -30,17 +30,10 @@ import androidx.compose.ui.window.DialogProperties
 import androidx.compose.ui.window.SecureFlagPolicy
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.ramcosta.composedestinations.annotation.Destination
+import com.ramcosta.composedestinations.annotation.ExternalModuleGraph
 import com.ramcosta.composedestinations.result.ResultBackNavigator
 import com.ramcosta.composedestinations.spec.DestinationStyle
 import java.util.concurrent.TimeUnit
-import net.mullvad.mullvadvpn.BuildConfig
-import net.mullvad.mullvadvpn.R
-import net.mullvad.mullvadvpn.compose.screen.MainGraph
-import net.mullvad.mullvadvpn.compose.state.VoucherDialogState
-import net.mullvad.mullvadvpn.compose.state.VoucherDialogUiState
-import net.mullvad.mullvadvpn.compose.util.MAX_VOUCHER_LENGTH
-import net.mullvad.mullvadvpn.compose.util.vouchersVisualTransformation
-import net.mullvad.mullvadvpn.constant.VOUCHER_LENGTH
 import net.mullvad.mullvadvpn.lib.model.DAYS_PER_VOUCHER_MONTH
 import net.mullvad.mullvadvpn.lib.model.RedeemVoucherError
 import net.mullvad.mullvadvpn.lib.ui.component.textfield.CustomTextField
@@ -50,7 +43,6 @@ import net.mullvad.mullvadvpn.lib.ui.designsystem.VariantButton
 import net.mullvad.mullvadvpn.lib.ui.tag.VOUCHER_INPUT_TEST_TAG
 import net.mullvad.mullvadvpn.lib.ui.theme.AppTheme
 import net.mullvad.mullvadvpn.lib.ui.theme.Dimens
-import net.mullvad.mullvadvpn.viewmodel.VoucherDialogViewModel
 import org.koin.androidx.compose.koinViewModel
 
 @Preview(device = Devices.TV_720p)
@@ -109,7 +101,7 @@ private fun PreviewRedeemVoucherDialogSuccess() {
     }
 }
 
-@Destination<MainGraph>(style = DestinationStyle.Dialog::class)
+@Destination<ExternalModuleGraph>(style = DestinationStyle.Dialog::class)
 @Composable
 fun RedeemVoucher(resultBackNavigator: ResultBackNavigator<Boolean>) {
     val vm = koinViewModel<VoucherDialogViewModel>()
