+++ android/lib/feature/redeemvoucher/impl/src/main/java/net/mullvad/mullvadvpn/feature/redeemvoucher/impl/VoucherDialogUiState.kt
@@ -1,4 +1,4 @@
-package net.mullvad.mullvadvpn.compose.state
+package net.mullvad.mullvadvpn.feature.redeemvoucher.impl
 
 import net.mullvad.mullvadvpn.lib.model.RedeemVoucherError
