+++ android/settings.gradle.kts
@@ -49,6 +49,7 @@ include(
     ":lib:feature:filter:impl",
     ":lib:feature:multihop:impl",
     ":lib:feature:problemreport:impl",
+    ":lib:feature:redeemvoucher:impl",
     ":lib:feature:serveripoverride:impl",
     ":lib:feature:splittunneling:impl",
     ":lib:map",
